Block editor: remove RootContainer
